Move elements out of sequences of names or fields when possible, instead of copying them.
This is tricky to detect. Elements are generally copied from container lvalues,
or if elements are const, while they are moved from container rvalues; but they
must be copied from view containers even if these containers are rvalues.

Move support for container conversions to a separate header `containers.h` to
avoid highlighting a private API at the beginning of `csv_record.h`.

PiperOrigin-RevId: 360163098