Do not include the current line number in a failure from `ReadCsvRecordFromString()`, or the current record index in `WriteCsvRecordToString()`.
These functions are commonly used in context with some higher level line
splitting/joining, which means that these numbers are not merely irrelevant but
misleading.

This forced changing the `standalone_record` parameter from a template parameter
to a member variable.

Also, clean up preconditions: `ReadCsvRecordFromString()` and
`WriteCsvRecordToString()` are not applicable to the mode with `CsvHeader`.

PiperOrigin-RevId: 360214815