Do not use an empty `CsvHeader` as a marker for no header. Explicitly mark whether a header is present.
This was already problematic in one case: `CsvReader::ReadRecord(CsvRecord&)`
should not be supported if reading the header was not requested, but should be
supported if reading the header failed. These cases used to be distinguished by
`healthy()`, which is a conservative approximation.

Another case emerged. If the code dynamically prepares a header, and writes
matching records, then it is weird if writing an empty header does not yield
an empty line, or if writing an empty record violates a precondition. In both
scenarios this was because `CsvWriter` mistakenly assumed that the mode with a
header is turned off.

Hence the changes:
* `CsvWriterBase::Options::set_header()` takes `absl::optional<CsvHeader>`.
* Add `Csv{Reader,Writer}Base::has_header()`, so that the precondition for
  supporting `CsvRecord` is easier to express and possible to check.
* The CSV format still does not support empty records. Writing an empty header
  is equivalent to writing a header with one empty field, just like writing an
  empty record in general. This is the minimum amount of surprise.

`Csv{Reader,Writer}Base::header()` continue to return `const CsvHeader&`, not
`absl::optional<CsvHeader>&`, because they are generally used in contexts where
the mode with a header is unconditionally turned on by surrounding code. There
is no need to bother the calling code with worrying whether conditions for the
returned optional to be engaged are satisfied.

PiperOrigin-RevId: 360634538