Do not bother with rvalue reference overloads of getters of various “members” in “containers”, primarily in `Options` classes.
While such rvalue reference overloads are sound, they are useful only in case of
a container having a single most important member. This is because they are
useful to enable move semantics in code like `MakeContainer().member()` or
`std::move(container).member()`, which ignores all members apart from one. It is
uncommon to write such code for a container with several equally important
members.

Similarly, and not coincidentally, there are rvalue reference overloads for
`absl::optional::operator*`, but not for `std::vector::operator[]`. Only
`std::get()` is peculiar; it has rvalue reference overloads, and it is used with
both kinds of containers: `std::variant` and `std::tuple`.

When moving a member out of a container with several members, if other members
are still needed, `std::move(container.member())` must be used instead of
`std::move(container).member()`, because `container` as a whole must remain
valid for accessing other members.

Only when moving out the single most important member (assuming that other
members are no longer needed, if they exist at all), rvalue reference verloads
of getters are useful, and `std::move(container).member()` is preferred over
`std::move(container.member())`, because it makes it easier for tools to track
which objects are moved from and should no longer be accessed.

The use case of having several members, and invalidating the whole object while
moving the last member, is not important enough to have its own API.

PiperOrigin-RevId: 360666868