Change deduction guides for `StringReader` when the constructor argument is a `std::string`. Instead of disabling deduction in this case, which was meant to avoid unexpected string copying:
* If the constructor argument is a `std::string` lvalue, deduce
  `StringReader<absl::string_view>`.

* Otherwise deduce the value type of the constructor argument as usual,
  i.e. deduce `StringReader<std::string>` for `std::string` rvalues.

This means that `StringReader(variable)` constructs a
`StringReader<absl::string_view>` no matter whether `variable` is an
`absl::string_view` or `std::string`. The code author likely does not want to
care about such difference, and does not want a string copy in either case.

Also, `StringReader(Function())` constructs a `StringReader<absl::string_view>`
no matter whether `Function()` returns an `absl::string_view` or
`const std::string&`.

OTOH `StringReader(std::move(variable))` for a `std::string` variable, and
`StringReader(Function())` for `Function()` returning a `std::string`,
construct a `StringReader<std::string>`. This is safe and efficient because
the string is moved, not copied.

Background for why this issue arose at all, and only for `StringReader`:

There is a particular limitation of C++: if we have a construct which aims to
store the value of an expression for later, whatever its type is, and the
expression has a reference type, then it is unclear whether the intent is to
store just the reference, or copy the dereferenced value. References are meant
to be transparent wrappers of values, which leads to the ambiguity.

This is why lambda captures specify “by copy” or “by reference” per variable.
This is why `std::ref()` and `std::cref()` exist: to disambiguate the intent for
certain constructs, like `std::bind()` and `std::make_tuple()`.

The problem does not arise if pointers are used instead of references. But using
a pointer instead of a reference also changes how it is later retrieved.
Changing references to pointers is not be applicable to `std::bind()` if the
underlying functor expects a reference and not a pointer, nor to
`std::make_tuple()` if we want a reference and not a pointer in the tuple.

Constructors of Riegeli classes do expect pointers if the object must outlive
the constructor. Hence in most cases deducing the intent from the type works
well: the value is stored with implicit dereferencing if it was a reference, but
the stored value will be a pointer if referring to an existing object is
intended.

`StringReader` is special, because for strings we have two common ways to pass a
string: `const std::string&` and `absl::string_view`. One is a reference, and
another is a value with a reference-like semantics, yet they are often used
interchangeably. If we have a `std::string`, either as a value or as a
reference, then `StringReader(&str)` makes it clear that `str` must outlive the
`StringReader` constructor. But if we have an `absl::string_view`, then passing
it by pointer would not make the lifetime requirement clear at all: the point is
not the lifetime of the `absl::string_view` itself, but of the character array
behind it, which is impossible to reflect by the convention of using pointers
for unusual lifetime requirements.

Hence with new rules `StringReader` will by default store a reference to the
string, because copying it is rarely necessary. The caller will need to ensure
that it is valid, similarly as copying an `absl::string_view` does not ensure
that the character array behind it is valid, and as capturing it by value in a
lambda is not sufficient, and as `std::make_tuple()` with `absl::string_view`
will not copy the character array either, whether `std::ref()` is used or not.
This applies no matter whether `absl::string_view` or `std::string` is used.

It is preferable to write `StringReader(&str)` if `str` is a `std::string` (no
matter whether a value or a reference), but `StringReader(str)` if `str` is an
`absl::string_view`, or if the caller does not want to care which is the case.
`StringReader<std::string>(str)` must be written for the rare case when copying
the string is intended.

* * *

Also, deduce `StringReader<absl::string_view>` for the constructor argument of
`const char*`, to ensure that conversion of `const char*` to `absl::string_view`
happens as early as possible, to compute `std::strlen()` at compile time if
possible. This lets `StringReader("foo")` compile at all (otherwise the
dependency was deduced as an array).

This change also fixes how deducing `StringReader<std::string>` was disabled.
If the constructor argument was a `std::string&`, then overload resolution
among deduction guides preferred to instantiate a template than to convert
`std::string&` to `const std::string&`, and deduction was not actually disabled.

Moral: avoid relying on ordering of overloads when templates are involved,
prefer writing the templates such that only the intended overloads match,
e.g. using SFINAE constraints.

Similarly, fix deduction guides for `Fd{,Stream,MMap}{Reader,Writer}` in case
the constructor argument is not `int` but something convertible.

PiperOrigin-RevId: 367189772