Let `Flush(FlushType::kFromObject)` apply to the immediate destination, and also to its dependent destination if that is owned.
In other words, `Flush(FlushType::kFromObject)` flushes the same objects that
`Close()` closes. Owned dependencies are meant to be treated as details of
the immediate `Writer` in hand, so that it is not needed to manipulate them
explicitly. `Flush(FlushType::kFromObject)` used to be defined inconsistently.
In particular:

* Compressors, `{Wrapped,Digesting,Limiting}Writer`, `RecordWriter` no longer
  propagate `Flush(FlushType::kFromObject)` to unowned destinations.

* `riegeli::tensorflow::FileWriter::Flush(FlushType::kFromObject)` now calls
  `::tensorflow::WritableFile::Flush()` for an owned destination.
  A `::tensorflow::WritableFile` can buffer data, so it should be flushed if
  it is owned.

* In Python `RecordWriter`, `flush(FlushType.FROM_OBJECT)` now calls `flush()`
  for an owned destination. Also, rename the constructor parameter of
  `Record{Reader,Writer}` from `close` to `owns_{src,dest}`, given that it
  no longer affects only closing.

Also, let `WriterOstream::flush()` call `Flush(FlushType::kFromProcess)`
instead of `Flush(FlushType::kFromObject)`. This matches the intent of
`std::ostream::flush()` better.

PiperOrigin-RevId: 367977361