In `Flush()`, let the default `flush_type` be `FlushType::kFromProcess`.
This flush mode ensures that in-process buffering does not prevent seeing
the output. It corresponds to `std::ostream::flush()`, `<< std::flush`,
`<< std::endl`, and `std::fflush()`.

Other flush types are intended for more special circumstances:
`FlushType::kFromObject` for flushing some layers out of a multilayered writer
in the same process, and `FlushType::kFromMachine` for making the data durable
even if the machine crashes soon.

Reword comments about `Flush()` and `FlushType` to highlight the intent rather
than the mechanism.

PiperOrigin-RevId: 368415805