Fix quadratic cost of repeated reallocations of a growing buffer.
This manifested in `ReadAll()` to `absl::string_view`, even though it tried
to mitigate this by calling `Pull()` with a recommended length of twice the
available length. This was not sufficient in case the source is a non-seekable
`FdReader` whose file descriptor provides at most some fixed length at a time.
Then, the buffer got filled until it was slightly more than half full. Each time
`BufferedReader::PullSlow()` tried to ask for space twice the available length,
slightly less than that was available in the current allocation, so the buffer
got reallocated, to a size only slightly larger than its current size. Reading
filled only a small part of the second half of the buffer, so each iteration
reallocated a growing buffer.

Here is the new design:

`ChainBlock::{Append,Prepend}Buffer()` is responsible for ensuring that the
capacity grows fast enough to avoid quadratic cost of repeated reallocations.

For `BufferedReader`:

1. It utilizes the allocated capacity of the buffer, instead of computing the
   exact buffer size on its own. This ensures that the buffer size grows fast
   enough to avoid quadratic cost.

2. It fills the rest of the existing buffer as long as there is any space in it,
   instead of asking for a new buffer if the current buffer is more than half
   full. This implies that twice as much as needed is allocated in the worst
   case rather than in every case.

3. It fills at most twice the intended buffer size. This constrains the effect
   of a single `Pull()` with a large mininum length on all future reads: the
   effective buffer size gets enlarged but not arbitrarily much.

4. Cosmetic changes in buffer length computations (take `size_hint` into
   account all the time, use `recommended_length` instead of `min_length` for a
   reallocated buffer).

Changes for `PullableReader` are similar to `BufferedReader` changes 1 and 3.
The filled buffer size does not grow geometrically but only to cover the next
fragment of the original source, but this is enough, because the allocation
grows geometrically.

Similar changes are applied to `riegeli::tensorflow::FileReader` and
`PushableWriter`.

`ReadAll()` and `ReadLine()` to `absl::string_view` no longer supply a
`recommended_length` argument to `Pull()`.

PiperOrigin-RevId: 368420589