Let the geometric growth of the destination of a `StringWriter` use the factor of 1.5 rather than 2. This is enough to ensure amortized fixed time of reallocation, and saves memory, at the cost of more allocations.
Do not bother with rounding the string capacity to an allocation size minus 1.
With a growth factor of 1.5 instead of 2 the simpler computation no longer
yields powers of 2, which would likely hit a bad case for a memory allocator
when single bytes for NUL terminators are added. In future `std::string` will be
able to itself adjust the capacity in a better way (after allocation can be more
exact than before): http://wg21.link/p0901, http://wg21.link/p0401.

Remove ensuring geometric growth from `Buffer::Reset()`. It does not preserve
existing contents, so the time cost of reallocation is constant in either case.

PiperOrigin-RevId: 368429460