Make `Reader::ReadAll(std::string&)` faster in the case without random access, by avoiding data copying and requesting larger fragments from the source.
The old implementation was to simply copy all fragments provided by `Pull()` to
the string.

The new implementation resizes the string to account for data known to be
available, and then fills all its remaining space, possibly by reading from the
source directly into the string and in one step.

PiperOrigin-RevId: 369182694