Fix inefficiency in handling a slowly growing `min_length` argument of `PullableReader::Pull()`.
The maximum length should be based on doubling the original
`recommended_length`, not doubling the final `recommended_length` after
subtracting the length already present in the buffer. The final
`recommended_length` might be 1 when the buffer grows by 1.

Also, do not pass `recommended_length` as `size_hint` when removing the used
prefix. `recommended_length` is valid only for the current call, while
`size_hint` hints about all future operations. If the buffer needs to grow at
this time, it should respect the geometric growth rate.

For `PushableWriter::Push()`, there is no need to constrain the buffer length by
doubling the `recommended_length`.

PiperOrigin-RevId: 370017401