Change the API of returning `RecordPosition`s: add `Record{Reader,Writer}::last_pos()` to replace the optional output parameter of `Record{Reader::Read,Writer::Write}Record()`.
In Python, add `last_pos` property to replace methods with names ending with
`_with_key` or `_with_keys`.

The old API will be removed after existing callers are resolved.

Reasons for doing this change:

* This makes the API surface smaller, at the cost of adding a constraint on the
  order of function calls (which was the original motivation of the coupling of
  getting the canonical position with reading/writing the record).

* There are plans to offer more ways for identifying records (custom keys,
  absolute record indices). Letting one particular way occupy the optional
  parameter of `{Read,Write}Record()` is harder to extend.

Also, change the terminology to use “canonical position” instead of “key”,
anticipating future changes with different notions of keys.

Cosmetic change in `RecordReader::CheckFileFormat()`: there is no need to ask
`ChunkReader` to try reading the next chunk if a chunk is already decoded, not
merely if some decoded records are waiting to be read.

PiperOrigin-RevId: 372612100