Add `OrderedVarint`: similar to `Varint` but preserves the ordering, and also allows to determine the encoded length from the first byte, but is limited by design to 64-bit integers.
Use it in the serialization of `RecordPosition` to/from bytes, so that the
encoding is shorter than the fixed width encoding used before, but still
preserves the ordering.

`RecordPosition::FromBytes()` temporarily understands both formats.
`RecordPosition` serialization is currently used only in
`RiegeliDatasetOp::Dataset::Iterator` serialization. Detecting the format by the
encoding length works in practice because the new format would yield an encoding
confusable with the old format only for insanely large values (in particular the
number of records in a chunk would have to be larger than `1 << 42`).

PiperOrigin-RevId: 374160408