Let `ReadOrderedVarint{32,64}()` accept only the canonical (shortest) encoding. Remove `ReadCanonicalOrderedVarint{32,64}()`.
Only the canonical encoding actually preserves the ordering as claimed.

While it is occasionally useful to be able to leave room for a varint of an
unknown magnitude and fill it later (which requires non-canonical encodings),
it is likely more important for encodings to be unique.

This departs from the varint convention.

Thanks Thomas Fischbacher for the suggestion!

PiperOrigin-RevId: 375897771