Fix an embarrassing design bug in `Writer::SeekSlow()`: the fast path of `Writer::Seek()` was incorrect, because data written so far after the new position are forgotten instead of being flushed.
Fortunately this was not a problem a practice, because random access for writing
likely has never been used at all with Riegeli.

(`Writer::Seek()` and `Writer::Size()` are meant for the paradigm of updating
pieces data in place, where the `Writer` is coupled with a `Reader` over the
same data, e.g. a file opened in read/write mode. These functions are so far
only supported by files. This paradigm was not utilized yet. OTOH
`Writer::Truncate()` belongs to a separate API layer and is not affected by this
issue.)

The fix is to remove that fast path of `Writer::Seek()`, replacing `SeekSlow()`
with `SeekImpl()` which has the same preconditions and responsibility as
`Seek()`.

This maintains the separation between the public method which is being called,
and the protected method which is being overridden. That separation is
occasionally useful (e.g. it allows adding postcondition checks in a central
place, it makes it easier to evolve the interface together with its usages).

PiperOrigin-RevId: 379258144