Reduce the timeout of an idle thread in `ThreadPool` from 60s to 1s.
The motivation for reusing threads across tasks is to make it cheaper to launch
a large number of short tasks, avoiding the overhead of thread creation, which
is a significant fraction of the total time only if tasks are short.

But if a thread is idle for a significant time, it means that the rate of task
creation is low, or that the number of tasks is decreasing. Conserving memory is
more important than reducing the overhead of thread creation by a tiny amount.

PiperOrigin-RevId: 380548490