Make `BufferedReader` easier to derive from, by reducing the need to define member functions with boilerplate code patterns:
Override `BufferedReader::{Done,SyncImpl}()` to automatically seek the source
back to the current position if not all buffered data were read. This is
feasible only if `SupportsRandomAccess()`.

Override `BufferedReader::SeekSlow()` to automatically take the buffer into
account. Derived classes can override `BufferedReader::SeekBehindBuffer()`
instead and not deal with synchronizing the buffer.

Remove `ClearBuffer()` from the protected interface.

Free buffer memory in `BufferedReader::Done()`.

Also, clarify the intent of `Reader::Sync()` that it should discard buffered
data, so that they are reread from the source if requested again (this is how
`BufferedReader::SyncImpl()` behaves). This makes `Sync()` suitable for ensuring
that interleaving reading from and writing to the same dependency works, if the
dependency supports that at all.

PiperOrigin-RevId: 381236079