When `Fd{,MMap}{Reader,Writer}` or `tensorflow::File{Reader,Writer}` is asked to open the file, and that fails, leave the `Reader` or `Writer` closed.
Also, when a `Reader` or `{,Backward}Writer` fails while closed, do not annotate
the `Status` with the current position.

As a consequence, failures to open the file are not annotated with the dummy
position of 0.

For consitency, remove an unnecessary assertion of `is_open()` in
`FailOperation()` also in `IstreamReader` and `OstreamWriter`.

PiperOrigin-RevId: 381246230