Change the `absl::StatusCode` from `kDataLoss` to `kInvalidArgument` for failed Riegeli/records parsing.
`kDataLossError` is meant for cases where data previously stored turned out
to be lost after all; `RecordReader` does not know whether this is the case.
`kInvalidArgument` is a better match if incoming data do not have the correct
format.

As a special case, failures which are considered extension points (unknown chunk
type and unknown compression method) return `kUnimplemented`.

PiperOrigin-RevId: 382302091