Remove `{,Backward}Writer::WriteHint()`.
In contrast to `Reader::ReadHint()`, its effects could be either a win (when
writing by copying into a buffer in either case: the buffer gets larger to
accommodate all data and thus there are fewer rounds of the low level writing
call) or a loss (when sharing a `Chain` or `Cord`: the larger buffer gets
dropped, possibly with extra rewriting of the buffer if it wasted memory; even
when writing from a single long `string_view` the data are unnecessarily copied
into a buffer instead of being written directly).

In any case the effect was small.

PiperOrigin-RevId: 384448514