Change the API of reading numeric values (bytes, varints, fixed-width integers) from returning `absl::optional` to setting an output parameter and returning `bool`.
It turned out that handling `absl::optional` results of reading functions is
sometimes inconvenient. That `absl::optional` result is almost always stored in
a local variable, and after immediate checking for possible failure, either
`*variable` is used in the following code, or `*variable` is copied to its
target location. In case of `bool` result, after immediate checking for possible
failure, either `variable` is used in the following code, or `variable` is
eliminated by reading the number directly to its target location. In both cases
the version with a `bool` result leads to simpler code.

The difference is especially striking when several values are read in sequence.
This is very annoying with `absl::optional` results, as it requires duplicating
the failure handling code, or creating a helper function which separates failure
handling from reading the sequence, or increasing the indentation level after
reading each value, or trying to read all values even if a previous operation
failed and then handling failures. With `bool` results this is easier: function
calls can be joined using `&&` or `||`, with a single failure handling code.

There are cases where `absl::optional` results are more convenient than `bool`
results, but they are rare. Primarily tests, which handle possible failure
together with examining the value, and which no longer need the value after it
is verified to match the expectation.

An argument for `absl::optional` results is that the API reflects the contract
more directly, in particular by not having to document (or guess) what is left
in the output parameter on failure, given that there is no output parameter.

An argument for `bool` results is consistency with similar functions for reading
strings, lines, records etc., where an output parameter is important for
performance. Functions with `bool` results also compose nicely, and for
sequences of values the reading code can be analogous to the writing code.

These functions with `bool` results are sometimes 10% faster than functions with
`absl::optional` results.

Considering all these arguments, I followed the “practicality beats purity”
motto and decided to migrate to `bool` results.

PiperOrigin-RevId: 394416382