Fix a race condition when abandoning an open `RecordWriter` which uses parallelism: let `RecordWriter::ParallelWorker::healthy()` lock the mutex.
The `ObjectState` of the `Worker` is normally modified either when background
work already finished or by `Fail()` called by the `Worker` itself (hence I did
not lock the mutex in `healthy()`, which is a protected member and thus not
called from outside), but abandoning an open `RecordWriter` is a case when
`Fail()` is called by the user of the `RecordWriter`.

Luckily this race condition was harmless in practice: reading a `bool` without
the lock might cause the `Worker` to see a stale value, which might cause it to
delay reacting to cancellation until the next time it takes the lock.

PiperOrigin-RevId: 400653129