Unconditionally support `Writer::Seek()` to the current position.
It turns out to be useful. Clients should not need to guard a `Seek()` with
checking whether the position changes if the `Writer` does not necessarily
support random access (and the position never changes in such case).

This is done by hardcoding the condition in `Seek()`, and changing `SeekImpl()`
to `SeekSlow()` which deals only with other cases.

Cosmetics: check `new_pos > start_pos()` instead of `new_pos >= start_pos()`
when equality is impossible anyway.
PiperOrigin-RevId: 406323472