Bring interfaces of compression dictionaries of Brotli, Zstd, and Zlib closer together.
Unify `Brotli{Reader,Writer}Base::Dictionaries` into `BrotliDictionary`,
suitable for either compression or decompression.

The cost is higher memory use of a `BrotliDictionary`, because the prepared
dictionary contains a copy of original data (except when it does not need to own
the original data then the copy was already being performed).

Unify `Zstd{Reader,Writer}Base::Dictionary` into `ZstdDictionary`, suitable for
either compression or decompression.

Change the API of setting the Zstd dictionary to provide both the contents and
the type in one call. This enables a simpler implementation which does not have
to care about a hypothetical case when the data is shared and then the type gets
changed in some of the objects.

Cosmetics: ensure that dictionaries are deleted in `Close()`.
PiperOrigin-RevId: 407075226