Reorder parameters of `BufferLength()` so that `recommended_length`, even though optional, is between `min_length` and `max_length`.
This is consistent with other functions, e.g. `Chain{,Block}::AppendBuffer()`.

Also, the nature of `max_length` is typically different in the other overload
when it is also `recommended_length`: it is really both, being some reasonably
small size, as opposed to potentially a huge limit in the overload with separate
parameters. Hence keeping `max_length` in the same position has less value.

PiperOrigin-RevId: 409915658