Add `{,Backward}Writer::Write{Chars,Bytes}()` which generalize `WriteZeros()` to an arbitrary character. This is useful e.g. for text formatting.
`WriteZeros()` remains as a separate function because it is the most common
usage. `WriteZerosSlow()` remains as the only virtual function because there are
little opportunities to optimize writing sequences of arbitrary characters for
particular `Writer` classes, compared to writing zeros (because we can have a
single statically allocated array of zeros).

PiperOrigin-RevId: 418145465