Add `Buffer::{Append,Prepend}SubstrTo()` to factor out common usages of `Buffer::ToCord()`.
This slightly changes underlying mechanisms: if a buffer needs to be copied for
appending to a `Cord` and it is between 512 and 4083 bytes, it is appended using
`Cord::Append(string_view)` instead of converting to a `Cord` and appending
using `Cord::Append(Cord)`.

Mark these functions, along with `Buffer::ToCord()`, as consuming `*this`,
because that was not clear from their names.

Rename `MakeFlatCord()` to `MakeBlockyCord()`. Move `{Append,Prepend}ToCord()`
from private functions in `chain.cc` to `base.h`, renaming them to
`{Append,Prepend}ToBlockyCord()`. These functions do not guarantee that new
pieces are flat, but they are meant for cases where pieces are large enough and
when overallocation is not needed, such as when converting from a `Chain` which
maintains block fitness invariants.

Cosmetic changes.

PiperOrigin-RevId: 425330123