If the `Manager` in `AnyDependency` is already the same `AnyDependency` type, avoid wrapping it again. Move the original contents instead.
This was already handled in case of a one-argument constructor, because the
appropriate explicit converting constructor is hidden by the move constructor if
the argument type matches.

Here it is handled in other cases: a constructor with an explicit `Manager`
type tag, constructing from a tuple of arguments, and `Reset()`.

These cases might arise from generic code. If a function or constructor template
accepts any suitable `Manager` type to store it in an `AnyDependency`, and the
caller happens to pass `AnyDependency` itself, then there is no need for extra
indirection, no matter how the stored `AnyDependency` is initialized.

PiperOrigin-RevId: 425854337