Coalesce overloads of functions taking a dependency which can bind to a reference (`ParseFromReader()`, `SerializeToWriter()`, `Snappy{Compress,Decompress}()`).
This simplifies code.

The cost is that this practically takes away the possibility to specify the
template argument explicitly, because that does not behave intuitively with
forwarding references. Hence this also disallows passing a tuple of constructor
arguments, because such a tuple needs the type to construct to be specified
separately. OTOH in these cases passing a tuple did not have any advantage over
passing a temporary, because the dependency is bound to the temporary without
moving it.

PiperOrigin-RevId: 426862794