Overhaul comments explaining `Dependency`.
Use comparison with `nullptr` as the idiomatic way to test whether a smart
pointer contains a value.

Change `kIsStable` from a function to a constant. The only reason why it was a
function was that before C++17 static constants require definitions if they are
ODR-used. In practice `kIsStable` is never ODR-used, and since C++17 this will
be moot.

Move the specialized part of `Dependency` to a `DependencyImpl` base class.
Individual specializations no longer need to provide `operator*` and
`operator->`, which are implemented by `Dependency` based on `get()`. They are
handled by `Dependency` itself, along with comparisons against `nullptr`.

In `AnyDependency`, make `Dependency::is_owning()` optional. It is not
applicable e.g. to `Dependency<absl::string_view, _>`, but it is useful to have
`AnyDependency<absl::string_view>`.

PiperOrigin-RevId: 427687431