Rename `healthy()` to `ok()`.
This matches `absl::Status` and `absl::StatusOr` with a matching semantics,
i.e. `ok() == status().ok()`.

One goal is to nudge callers to checking `ok()` instead of `status()`,
reassuring them that the property is familiar.

The name `healthy()` predates integration with `Status` and I thought it was
easier to use as an adjective in prose, but it looks like `ok()` is fine (“OK”
in prose).

PiperOrigin-RevId: 428474643