Define a converting `operator=` where an implicit conversion exists, and performing the converting assignment using an implicit conversion and then regular assignment might be more expensive.
An implicit conversion indicates that it is idiomatic to write just `x = y`
instead of e.g. `x.Reset(y)`, even if the types do not match exactly. This can
be implemented using one move instead of two.

There is no need to do this for types which are always cheap to move.

PiperOrigin-RevId: 429275302