Pass functors by forwarding references instead of by value. Forward or call them using `std::forward<F>(f)(args)` if they are called only once, but call them using `f(args)` if they are called multiple times.
This avoids copying them. For mutable functors, changes to their state is
visible to the caller. Also, functors callable only once (with `operator() &&`)
are permitted in cases when they are called only once, but they are not
permitted if they are called multiple times.

This matches modern conventions in C++, e.g.
https://en.cppreference.com/w/cpp/utility/optional/transform (they are even more
general by using `std::invoke`, but this requires C++17.)

PiperOrigin-RevId: 429279137