Add standalone functions `riegeli::Read{,AndAppend}All()`, `riegeli::Write()`, and `riegeli::CopyAll()` which combine creating a `Reader` and/or `Writer`, reading/writing/copying a single piece, and closing owned dependencies.
They are not that useful in Riegeli itself, but they capture a common pattern in
other projects. It is weird that only Snappy compression provides that shortcut
(where it was motivated by efficiency, as it eliminates emulation of streaming
compression/decompression which is not supported by the Snappy engine) but not
other compressors/decompressors.

PiperOrigin-RevId: 429300592