Move `DependencyImpl` specializations to the same header as some template argument. If all template arguments are external types, they are moved to `dependency.h`.
Similarly for `AnyDependencyTraits`, moved to `any_dependency.h`.

This avoids the need to include particular headers just to see specializations,
which is prone to violating the ODR. While these specializations were needed for
particular `Reader` and `Writer` classes (which did include the relevant
headers), nobody says that they have to be used only by these classes.

PiperOrigin-RevId: 429307835