Add `growing_source` option to file reading. By default reading from a growing source is not supported, which means that the `Reader` can cache the file size.
This allows to optimize various scenarios, including repeated seeking forwards
(which checks the target position against the size), `SupportsSize()` followed
by `Size()`, `ReadAll()` followed by `VerifyEnd()`, and readng an unknown amount
from `std::istream` or `FILE*`.

Even if the file grew asynchronously between checking the size and reading,
`riegeli::ReadAll()` (which calls `Reader::ReadAll()` and `VerifyEnd()`) should
return a snapshot of the file, not fail. `VerifyEnd()` is meant to fail when the
contents are invalid, not when timing was unfortunate.

Perform another `VerifyEnd()` optimization: if it is delegated to the source,
e.g. as in `DigestingReader`, perform the verification only in the source, and
propagate any failure by `MakeBuffer()`. Calling both `Reader::VerifyEnd()` and
delegating to the source leads to exponential explosion of the number of checks
in deeply nested `Reader `stacks with owned sources.

PiperOrigin-RevId: 432911562