Add `CsvReaderBase& self` parameter to the recovery function.
Having access to the `CsvReader` is useful to e.g. get the current line number,
and it is not always convenient to access it otherwise on the caller side,
because the recovery function is usually passed to the initializer of the
`CsvReader`, and you cannot refer to a variable from its own initializer.

PiperOrigin-RevId: 439535649